restructuring graphql code a bit
